move favorites related steps to own context
