+++ tests/acceptance/features/bootstrap/WebDav.php
@@ -1343,57 +1343,6 @@ trait WebDav {
 	}
 
 	/**
-	 * @Then /^as user "([^"]*)" the (?:file|folder|entry) "([^"]*)" should be favorited$/
-	 *
-	 * @param string $user
-	 * @param string $path
-	 * @param integer $expectedValue 0|1
-	 *
-	 * @return void
-	 */
-	public function asUserTheFileOrFolderShouldBeFavorited($user, $path, $expectedValue = 1) {
-		$property = "{http://owncloud.org/ns}favorite";
-		$this->asUserFolderShouldContainAPropertyWithValue(
-			$user, $path, $property, $expectedValue
-		);
-	}
-
-	/**
-	 * @Then /^as user "([^"]*)" (?:file|folder|entry) "([^"]*)" should not be favorited$/
-	 *
-	 * @param string $user
-	 * @param string $path
-	 *
-	 * @return void
-	 */
-	public function asUserFileShouldNotBeFavorited($user, $path) {
-		$this->asUserTheFileOrFolderShouldBeFavorited($user, $path, 0);
-	}
-
-	/**
-	 * @Then /^as the user (?:file|folder|entry) "([^"]*)" should be favorited$/
-	 *
-	 * @param string $path
-	 * @param integer $expectedValue 0|1
-	 *
-	 * @return void
-	 */
-	public function asTheUserFileOrFolderShouldBeFavorited($path, $expectedValue = 1) {
-		$this->asUserTheFileOrFolderShouldBeFavorited($this->getCurrentUser(), $path, $expectedValue);
-	}
-
-	/**
-	 * @Then /^as the user (?:file|folder|entry) "([^"]*)" should not be favorited$/
-	 *
-	 * @param string $path
-	 *
-	 * @return void
-	 */
-	public function asTheUserFileOrFolderShouldNotBeFavorited($path) {
-		$this->asTheUserFileOrFolderShouldBeFavorited($path, 0);
-	}
-
-	/**
 	 * @Then the single response should contain a property :key with value like :regex
 	 *
 	 * @param string $key
@@ -2482,95 +2431,6 @@ trait WebDav {
 	}
 
 	/**
-	 * @When user :user favorites element :path using the WebDAV API
-	 * @Given user :user has favorited element :path
-	 *
-	 * @param string $user
-	 * @param string $path
-	 *
-	 * @return void
-	 * @throws \Sabre\HTTP\ClientException
-	 * @throws \Sabre\HTTP\ClientHttpException
-	 */
-	public function userFavoritesElement($user, $path) {
-		$this->response = $this->changeFavStateOfAnElement(
-			$user, $path, 1
-		);
-	}
-
-	/**
-	 * @When the user favorites element :path using the WebDAV API
-	 * @Given the user has favorited element :path
-	 *
-	 * @param string $path
-	 *
-	 * @return void
-	 * @throws \Sabre\HTTP\ClientException
-	 * @throws \Sabre\HTTP\ClientHttpException
-	 */
-	public function theUserFavoritesElement($path) {
-		$this->response = $this->changeFavStateOfAnElement(
-			$this->getCurrentUser(), $path, 1
-		);
-	}
-
-	/**
-	 * @When user :user unfavorites element :path using the WebDAV API
-	 * @Given user :user has unfavorited element :path
-	 *
-	 * @param string $user
-	 * @param string $path
-	 *
-	 * @return void
-	 * @throws \Sabre\HTTP\ClientException
-	 * @throws \Sabre\HTTP\ClientHttpException
-	 */
-	public function userUnfavoritesElement($user, $path) {
-		$this->response = $this->changeFavStateOfAnElement(
-			$user, $path, 0
-		);
-	}
-
-	/**
-	 * @When the user unfavorites element :path using the WebDAV API
-	 * @Given the user has unfavorited element :path
-	 *
-	 * @param string $path
-	 *
-	 * @return void
-	 * @throws \Sabre\HTTP\ClientException
-	 * @throws \Sabre\HTTP\ClientHttpException
-	 */
-	public function theUserUnfavoritesElement($path) {
-		$this->response = $this->changeFavStateOfAnElement(
-			$this->getCurrentUser(), $path, 0
-		);
-	}
-
-	/**
-	 * Set the elements of a proppatch
-	 *
-	 * @param string $user
-	 * @param string $path
-	 * @param int $favOrUnfav 1 = favorite, 0 = unfavorite
-	 *
-	 * @return bool
-	 */
-	public function changeFavStateOfAnElement(
-		$user, $path, $favOrUnfav
-	) {
-		$client = $this->getSabreClient($user);
-		$properties = [
-			'{http://owncloud.org/ns}favorite' => $favOrUnfav
-		];
-
-		$response = $client->proppatch(
-			$this->getDavFilesPath($user) . $path, $properties
-		);
-		return $response;
-	}
-
-	/**
 	 * @When user :user stores etag of element :path using the WebDAV API
 	 * @Given user :user has stored etag of element :path
 	 *
@@ -2715,86 +2575,6 @@ trait WebDav {
 				"'$expectedHeaderValue' does not match '$returnedHeader'"
 			);
 		}
-	}
-
-	/**
-	 * @Then /^user "([^"]*)" in folder "([^"]*)" should (not|)\s?have favorited the following elements$/
-	 *
-	 * @param string $user
-	 * @param string $folder
-	 * @param string $shouldOrNot (not|)
-	 * @param TableNode $expectedElements
-	 *
-	 * @return void
-	 */
-	public function checkFavoritedElements(
-		$user, $folder, $shouldOrNot, $expectedElements
-	) {
-		$this->userListsFavoriteOfFolder($user, $folder, null);
-		$this->propfindResultShouldContainEntries(
-			$user, $shouldOrNot, $expectedElements
-		);
-	}
-
-	/**
-	 * @Then /^the user in folder "([^"]*)" should (not|)\s?have favorited the following elements$/
-	 *
-	 * @param string $folder
-	 * @param string $shouldOrNot (not|)
-	 * @param TableNode $expectedElements
-	 *
-	 * @return void
-	 */
-	public function checkFavoritedElementsForCurrentUser(
-		$folder, $shouldOrNot, $expectedElements
-	) {
-		$this->checkFavoritedElements(
-			$this->getCurrentUser(), $folder, $shouldOrNot, $expectedElements
-		);
-	}
-
-	/**
-	 * @When /^user "([^"]*)" lists the favorites of folder "([^"]*)" and limits the result to ([\d*]) elements using the WebDAV API$/
-	 *
-	 * @param string $user
-	 * @param string $folder
-	 * @param int $limit
-	 *
-	 * @return void
-	 */
-	public function userListsFavoriteOfFolder($user, $folder, $limit = null) {
-		$baseUrl = $this->getBaseUrl();
-		$password = $this->getPasswordForUser($user);
-		$body
-			= "<?xml version='1.0' encoding='utf-8' ?>\n" .
-			"	<oc:filter-files xmlns:a='DAV:' xmlns:oc='http://owncloud.org/ns' >\n" .
-			"		<a:prop><oc:favorite/></a:prop>\n" .
-			"		<oc:filter-rules><oc:favorite>1</oc:favorite></oc:filter-rules>\n";
-		
-		if ($limit !== null) {
-			$body .= "		<oc:search>\n" .
-					 "			<oc:limit>$limit</oc:limit>\n" .
-					 "		</oc:search>\n";
-		}
-
-		$body .= "	</oc:filter-files>";
-		$response = WebDavHelper::makeDavRequest(
-			$baseUrl, $user, $password, "REPORT", "/", null, $body, null,
-			$this->getDavPathVersion()
-		);
-		$this->setResponse($response);
-	}
-	
-	/**
-	 * @When /^the user lists the favorites of folder "([^"]*)" and limits the result to ([\d*]) elements using the WebDAV API$/
-	 *
-	 * @param string $folder
-	 * @param int $limit
-	 *
-	 * @return void
-	 */
-	public function listFavoriteOfFolder($folder, $limit = null) {
-		$this->userListsFavoriteOfFolder($this->getCurrentUser(), $folder, $limit);
 	}
 
 	/**
